+++ src/Exceptions/PermissionDoesNotExist.php
@@ -2,8 +2,12 @@
 
 namespace Spatie\Permission\Exceptions;
 
-use Exception;
+use InvalidArgumentException;
 
-class PermissionDoesNotExist extends Exception
+class PermissionDoesNotExist extends InvalidArgumentException
 {
+    public static function create(string $permissionName, string $guardName = '')
+    {
+        return new static("There is no permission named `{$permissionName}` for guard `{$guardName}`.");
+    }
 }