+++ src/Exceptions/RoleDoesNotExist.php
@@ -2,8 +2,12 @@
 
 namespace Spatie\Permission\Exceptions;
 
-use Exception;
+use InvalidArgumentException;
 
-class RoleDoesNotExist extends Exception
+class RoleDoesNotExist extends InvalidArgumentException
 {
+    public static function create(string $roleName)
+    {
+        return new static("There is no role named `{$roleName}`.");
+    }
 }