+++ src/Models/Role.php
@@ -5,74 +5,83 @@ namespace Spatie\Permission\Models;
 use Illuminate\Database\Eloquent\Model;
 use Spatie\Permission\Traits\HasPermissions;
 use Spatie\Permission\Exceptions\RoleDoesNotExist;
+use Spatie\Permission\Exceptions\GuardDoesNotMatch;
+use Spatie\Permission\Exceptions\RoleAlreadyExists;
 use Spatie\Permission\Contracts\Role as RoleContract;
 use Spatie\Permission\Traits\RefreshesPermissionCache;
+use Illuminate\Database\Eloquent\Relations\MorphToMany;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 
 class Role extends Model implements RoleContract
 {
     use HasPermissions;
     use RefreshesPermissionCache;
 
-    /**
-     * The attributes that aren't mass assignable.
-     *
-     * @var array
-     */
     public $guarded = ['id'];
 
-    /**
-     * Create a new Eloquent model instance.
-     *
-     * @param array $attributes
-     */
     public function __construct(array $attributes = [])
     {
+        $attributes['guard_name'] = $attributes['guard_name'] ?? config('auth.defaults.guard');
+
         parent::__construct($attributes);
 
-        $this->setTable(config('laravel-permission.table_names.roles'));
+        $this->setTable(config('permission.table_names.roles'));
+    }
+
+    public static function create(array $attributes = [])
+    {
+        $attributes['guard_name'] = $attributes['guard_name'] ?? config('auth.defaults.guard');
+
+        if (static::where('name', $attributes['name'])->where('guard_name', $attributes['guard_name'])->first()) {
+            throw RoleAlreadyExists::create($attributes['name'], $attributes['guard_name']);
+        }
+
+        return static::query()->create($attributes);
     }
 
     /**
      * A role may be given various permissions.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
      */
-    public function permissions()
+    public function permissions(): BelongsToMany
     {
         return $this->belongsToMany(
-            config('laravel-permission.models.permission'),
-            config('laravel-permission.table_names.role_has_permissions')
+            config('permission.models.permission'),
+            config('permission.table_names.role_has_permissions')
         );
     }
 
     /**
-     * A role may be assigned to various users.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
+     * A role belongs to some users of the model associated with its guard.
      */
-    public function users()
+    public function users(): MorphToMany
     {
-        return $this->belongsToMany(
-            config('auth.model') ?: config('auth.providers.users.model'),
-            config('laravel-permission.table_names.user_has_roles')
+        return $this->morphedByMany(
+            getModelForGuard($this->attributes['guard_name']),
+            'model',
+            config('permission.table_names.model_has_roles'),
+            'role_id',
+            'model_id'
         );
     }
 
     /**
-     * Find a role by its name.
+     * Find a role by its name and guard name.
      *
      * @param string $name
+     * @param string|null $guardName
      *
-     * @throws RoleDoesNotExist
+     * @return \Spatie\Permission\Contracts\Role|\Spatie\Permission\Models\Role
      *
-     * @return Role
+     * @throws \Spatie\Permission\Exceptions\RoleDoesNotExist
      */
-    public static function findByName($name)
+    public static function findByName(string $name, $guardName = null): RoleContract
     {
-        $role = static::where('name', $name)->first();
+        $guardName = $guardName ?? config('auth.defaults.guard');
+
+        $role = static::where('name', $name)->where('guard_name', $guardName)->first();
 
         if (! $role) {
-            throw new RoleDoesNotExist();
+            throw RoleDoesNotExist::create($name);
         }
 
         return $role;
@@ -84,11 +93,17 @@ class Role extends Model implements RoleContract
      * @param string|Permission $permission
      *
      * @return bool
+     *
+     * @throws \Spatie\Permission\Exceptions\GuardMismatch
      */
-    public function hasPermissionTo($permission)
+    public function hasPermissionTo($permission): bool
     {
         if (is_string($permission)) {
-            $permission = app(Permission::class)->findByName($permission);
+            $permission = app(Permission::class)->findByName($permission, $this->getDefaultGuardName());
+        }
+
+        if (! $this->getGuardNames()->contains($permission->guard_name)) {
+            throw GuardDoesNotMatch::create($permission->guard_name, $this->getGuardNames());
         }
 
         return $this->permissions->contains('id', $permission->id);