+++ src/Traits/HasPermissions.php
@@ -2,16 +2,19 @@
 
 namespace Spatie\Permission\Traits;
 
+use Illuminate\Support\Collection;
+use Spatie\Permission\PermissionRegistrar;
 use Spatie\Permission\Contracts\Permission;
+use Spatie\Permission\Exceptions\GuardDoesNotMatch;
 
 trait HasPermissions
 {
     /**
      * Grant the given permission(s) to a role.
      *
-     * @param string|array|Permission|\Illuminate\Support\Collection $permissions
+     * @param string|array|\Spatie\Permission\Contracts\Permission|\Illuminate\Support\Collection $permissions
      *
-     * @return HasPermissions
+     * @return $this
      */
     public function givePermissionTo(...$permissions)
     {
@@ -19,6 +22,9 @@ trait HasPermissions
             ->flatten()
             ->map(function ($permission) {
                 return $this->getStoredPermission($permission);
+            })
+            ->each(function ($permission) {
+                $this->ensureModelSharesGuard($permission);
             })
             ->all();
 
@@ -32,7 +38,7 @@ trait HasPermissions
     /**
      * Remove all current permissions and set the given ones.
      *
-     * @param array ...$permissions
+     * @param string|array|\Spatie\Permission\Contracts\Permission|\Illuminate\Support\Collection $permissions
      *
      * @return $this
      */
@@ -46,9 +52,9 @@ trait HasPermissions
     /**
      * Revoke the given permission.
      *
-     * @param $permission
+     * @param \Spatie\Permission\Contracts\Permission|string $permission
      *
-     * @return HasPermissions
+     * @return $this
      */
     public function revokePermissionTo($permission)
     {
@@ -60,20 +66,66 @@ trait HasPermissions
     }
 
     /**
-     * @param string|array|Permission|\Illuminate\Support\Collection $permissions
+     * @param string|array|\Spatie\Permission\Contracts\Permission|\Illuminate\Support\Collection $permissions
      *
-     * @return Permission
+     * @return \Spatie\Permission\Contracts\Permission
      */
-    protected function getStoredPermission($permissions)
+    protected function getStoredPermission($permissions): Permission
     {
         if (is_string($permissions)) {
-            return app(Permission::class)->findByName($permissions);
+            return app(Permission::class)->findByName($permissions, $this->getDefaultGuardName());
         }
 
         if (is_array($permissions)) {
-            return app(Permission::class)->whereIn('name', $permissions)->get();
+            return app(Permission::class)
+                ->whereIn('name', $permissions)
+                ->whereId('guard_name', $this->getGuardNames())
+                ->get();
         }
 
         return $permissions;
+    }
+
+    /**
+     * @param \Spatie\Permission\Contracts\Permission|\Spatie\Permission\Contracts\Role $roleOrPermission
+     *
+     * @throws \Spatie\Permission\Exceptions\GuardMismatch
+     */
+    protected function ensureModelSharesGuard($roleOrPermission)
+    {
+        if (! $this->getGuardNames()->contains($roleOrPermission->guard_name)) {
+            throw GuardDoesNotMatch::create($roleOrPermission->guard_name, $this->getGuardNames());
+        }
+    }
+
+    protected function getGuardNames(): Collection
+    {
+        if ($this->guard_name) {
+            return collect($this->guard_name);
+        }
+
+        return collect(config('auth.guards'))
+            ->map(function ($guard) {
+                return config("auth.providers.{$guard['provider']}.model");
+            })
+            ->filter(function ($model) {
+                return get_class($this) === $model;
+            })
+            ->keys();
+    }
+
+    protected function getDefaultGuardName(): string
+    {
+        $default = config('auth.defaults.guard');
+
+        return $this->getGuardNames()->first() ?: $default;
+    }
+
+    /**
+     * Forget the cached permissions.
+     */
+    public function forgetCachedPermissions()
+    {
+        app(PermissionRegistrar::class)->forgetCachedPermissions();
     }
 }