+++ src/Traits/RefreshesPermissionCache.php
@@ -2,30 +2,23 @@
 
 namespace Spatie\Permission\Traits;
 
+use Illuminate\Database\Eloquent\Model;
 use Spatie\Permission\PermissionRegistrar;
 
 trait RefreshesPermissionCache
 {
     public static function bootRefreshesPermissionCache()
     {
-        static::created(function ($model) {
-            $model->forgetCachedPermissions();
+        static::created(function (Model $model) {
+            app(PermissionRegistrar::class)->forgetCachedPermissions();
         });
 
-        static::updated(function ($model) {
-            $model->forgetCachedPermissions();
+        static::updated(function (Model $model) {
+            app(PermissionRegistrar::class)->forgetCachedPermissions();
         });
 
-        static::deleted(function ($model) {
-            $model->forgetCachedPermissions();
+        static::deleted(function (Model $model) {
+            app(PermissionRegistrar::class)->forgetCachedPermissions();
         });
-    }
-
-    /**
-     *  Forget the cached permissions.
-     */
-    public function forgetCachedPermissions()
-    {
-        app(PermissionRegistrar::class)->forgetCachedPermissions();
     }
 }