+++ CHANGELOG.md
@@ -1,6 +1,64 @@
 # Changelog
 
-All Notable changes to `laravel-permission` will be documented in this file
+All notable changes to `laravel-permission` will be documented in this file
+
+## 2.2.0 - 2017-07-01
+- `hasanyrole` and `hasallrole` can accept multiple roles
+
+## 2.1.6 - 2017-06-06
+- fixed a bug where `hasPermissionTo` wouldn't use the right guard name
+
+## 2.1.5 - 2017-05-17
+- fixed a bug that didn't allow you to assign a role or permission when using multiple guards
+
+## 2.1.4 - 2017-05-10
+
+- add `model_type` to the primary key of tables that use a polymorphic relationship
+
+## 2.1.3 - 2017-04-21
+
+- fixed a bug where the role()/permission() relation to user models would be saved incorrectly
+- added users() relation on Permission and Role
+
+## 2.1.2 - 2017-04-20
+
+- fix a bug where the `role()`/`permission()` relation to user models would be saved incorrectly
+- add `users()` relation on `Permission` and `Role`
+
+## 2.0.2 - 2017-04-13
+
+- check for duplicates when adding new roles and permissions
+
+## 2.0.1 - 2017-04-11
+
+- fix the order of the `foreignKey` and `relatedKey` in the relations
+
+## 2.0.0 - 2017-04-10
+
+- cache expiration is now configurable and set to one day by default
+- roles and permissions can now be assigned to any model through the `HasRoles` trait
+- removed deprecated `hasPermission` method
+- renamed config file from `laravel-permission` to `permission`.
+
+## 1.12.0
+
+- made foreign key name to users table configurable
+
+## 1.11.1
+
+- `hasPermissionTo` uses the cache to avoid extra queries when it is called multiple times
+
+## 1.11.0
+
+- add `getDirectPermissions`, `getPermissionsViaRoles`, `getAllPermissions`
+
+## 1.10.0 - 2017-02-22
+
+- add `hasAnyPermission`
+
+## 1.9.0 - 2017-02-20
+- add `log_registration_exception` in settings file
+- fix for ambiguous column name `id` when using the role scope 
 
 ## 1.8.0 - 2017-02-09
 - `hasDirectPermission` method is now public